Initial commit
This project essentially combined elements from the
metalsmith-webpack-suite and the Vue.js webpack-simple template.
It's a static site with a small dynamic component in the home page
Vue.js based pager.
The pager is designed to work with a dataset that is
a) small enough to get the whole thing in one http request
   and hold in memory
but
b) large enough or complex enough such that you don't want to
   display the whole thing

For a personal photo travel site, it will be quite a while before
the number of images exceeds assumption a) above.
Paging the images a subset at a time reduces the number of image files
that need to be transferred over the wire.

The initial site design simply copied my portfolio at
michaelscottgraham.com, and changed the color scheme. This was done
to speed release and get something out the door. This initial
version was completed over the course of a few weeknights and
half a Saturday.